refactor: re-organize structure
