+++ memote/suite/cli/callbacks.py
